requirements.txt
google-genai>=0.8.0
groq>=0.11.0
playwright>=1.40.0
sounddevice>=0.4.6
scipy>=1.11.0
numpy>=1.26.0
pyttsx3>=2.90
python-dotenv>=1.0.0
